fixed major Photons.merge_orders() bug
The code did not consistently handle the exposure time and the "adding" of events for different orders. This resulted in absolute fluxes that were essentially the addition of the fluxes from individual orders (sometimes more due to exposure time issues). With the fix, the effective area associated with photons is adjusted according to overlap in orders and the exposure time is held fixed.